+++ src/game.rs
@@ -10,18 +10,19 @@ pub struct Game {
     pile: Pile,
     score: u16, // TODO Max of 65,535 is enough? (depends on ACCELERATION_SCORE_POINTS and maybe on ACCELERATION_FACTOR)
     current_tick_duration: Duration,
+    pub is_column_locked: bool,
     rng: fastrand::Rng,
 }
 
 impl Game {
     pub const BOARD_WIDTH: u8 = 6;
-    pub const BOARD_HEIGHT: i8 = 13;
+    pub const BOARD_HEIGHT: u8 = 13;
 
     const INITIAL_TICK_DURATION: Duration = Duration::from_millis(750);
     const MIN_TICK_DURATION: Duration = Duration::from_millis(100);
 
-    const ACCELERATION_SCORE_POINTS: u16 = 100; // New score points required for acceleration
-    const ACCELERATION_FACTOR: f64 = 0.98; // Reduce the current tick duration by 2%
+    const ACCELERATION_SCORE_POINTS: u16 = 100; // new score points required for acceleration
+    const ACCELERATION_FACTOR: f64 = 0.98; // reduce the current tick duration by 2%
 
     const FALLING_COLUMN_NOT_INITIALIZED_ERROR: &str = "Falling column must be initialized";
 
@@ -31,41 +32,58 @@ impl Game {
         Ok(Self {
             column_next: Self::create_column(&mut rng),
             column_falling: None,
-            pile: Pile::new(Self::BOARD_WIDTH, Self::BOARD_HEIGHT as u8),
+            pile: Pile::new(Self::BOARD_WIDTH, Self::BOARD_HEIGHT),
             score: 0,
             current_tick_duration: Self::INITIAL_TICK_DURATION,
+            is_column_locked: false,
             rng,
         })
     }
 
-    pub fn start(&mut self) {
-        self.transition_next_column_to_falling();
+    pub fn start(&mut self) -> anyhow::Result<()> {
         self.pile.clear();
         self.score = 0;
         self.current_tick_duration = Self::INITIAL_TICK_DURATION;
+        self.is_column_locked = false;
+        self.transition_next_column_to_falling();
+        self.rng = create_rng()?;
+
+        Ok(())
     }
 
     pub fn tick(&mut self) -> bool {
+        if self.pile.is_overflowed() {
+            self.pile.lock_final_gem();
+            return false;
+        }
+
         let column_falling = self.column_falling.as_mut().expect(Self::FALLING_COLUMN_NOT_INITIALIZED_ERROR);
 
-        let next_y_bottom = column_falling.y_bottom() + 1;
         let x = column_falling.x();
 
-        if next_y_bottom < Self::BOARD_HEIGHT && !self.pile.is_occupied(x, next_y_bottom) {
+        let is_column_descended_one_step = if self.pile.will_next_position_fit(x, column_falling) {
             column_falling.move_down(1);
+            true
         } else {
-            let is_locked = self.pile.lock(column_falling);
-            if !is_locked {
+            false
+        };
+
+        if !is_column_descended_one_step || !self.pile.will_next_position_fit(x, column_falling) {
+            self.pile.lock(column_falling);
+
+            if self.pile.is_overflowed() {
+                self.pile.lock_final_gem();
                 return false;
             }
 
+            self.is_column_locked = true;
             self.transition_next_column_to_falling();
-            self.tick();
         }
 
         if self.score > 0 && self.score.is_multiple_of(Self::ACCELERATION_SCORE_POINTS) {
             self.accelerate(Self::ACCELERATION_FACTOR);
         }
+
         true
     }
 
@@ -73,10 +91,7 @@ impl Game {
         let column_falling = self.column_falling.as_mut().expect(Self::FALLING_COLUMN_NOT_INITIALIZED_ERROR);
 
         let x = column_falling.x();
-        let y_bottom = column_falling.y_bottom();
-
-        // If the bottom Gem can move left, the whole column can move left
-        if x > 0 && !self.pile.is_occupied(x - 1, y_bottom) {
+        if x > 0 && self.pile.will_next_position_fit(x - 1, column_falling) {
             column_falling.move_left();
         }
     }
@@ -85,40 +100,37 @@ impl Game {
         let column_falling = self.column_falling.as_mut().expect(Self::FALLING_COLUMN_NOT_INITIALIZED_ERROR);
 
         let x = column_falling.x();
-        let y_bottom = column_falling.y_bottom();
-
-        // If the bottom Gem can move right, the whole column can move right
-        if x < Self::BOARD_WIDTH - 1 && !self.pile.is_occupied(x + 1, y_bottom) {
+        if x < Self::BOARD_WIDTH - 1 && self.pile.will_next_position_fit(x + 1, column_falling) {
             column_falling.move_right();
         }
     }
 
     pub const fn rotate_up(&mut self) {
         let column_falling = self.column_falling.as_mut().expect(Self::FALLING_COLUMN_NOT_INITIALIZED_ERROR);
+
         column_falling.rotate_up();
     }
 
     pub const fn rotate_down(&mut self) {
         let column_falling = self.column_falling.as_mut().expect(Self::FALLING_COLUMN_NOT_INITIALIZED_ERROR);
+
         column_falling.rotate_down();
     }
 
     pub fn drop(&mut self) {
         let column_falling = self.column_falling.as_mut().expect(Self::FALLING_COLUMN_NOT_INITIALIZED_ERROR);
+
         let x = column_falling.x();
+        let first_occupied_y = (0..Self::BOARD_HEIGHT).find(|&y| self.pile.get(x, y).is_some()).unwrap_or(Self::BOARD_HEIGHT);
+        let first_occupied_y = i8::try_from(first_occupied_y).expect("Board height should fit in `i8`");
 
-        // Find the first occupied row in the target column (top-down)
-        let first_occupied_y = (0..Self::BOARD_HEIGHT).find(|&y| self.pile.is_occupied(x, y)).unwrap_or(Self::BOARD_HEIGHT);
-
-        // Target is the row immediately above the first occupied Gem (or the floor)
-        let distance = (first_occupied_y - 1) - column_falling.y_bottom();
+        let distance = first_occupied_y - 1 - column_falling.y_bottom();
         if distance > 0 {
             column_falling.move_down(distance);
         }
-        self.tick();
     }
 
-    pub const fn _get_next_column(&self) -> &Column {
+    pub const fn get_next_column(&self) -> &Column {
         &self.column_next
     }
 
@@ -131,16 +143,20 @@ impl Game {
     }
 
     pub const fn tick_rate(&self) -> Duration {
-        self.current_tick_duration
+        if self.is_column_locked { Duration::ZERO } else { self.current_tick_duration }
+        //self.current_tick_duration
     }
 
     fn create_column(rng: &mut fastrand::Rng) -> Column {
-        let x = rng.u8(..Self::BOARD_WIDTH);
-        Column::new(x, -3, rng)
+        Column::new(0, 0, rng)
     }
 
     fn transition_next_column_to_falling(&mut self) {
-        let column_next = std::mem::replace(&mut self.column_next, Self::create_column(&mut self.rng));
+        let x = self.rng.u8(..Self::BOARD_WIDTH);
+
+        let mut column_next = std::mem::replace(&mut self.column_next, Self::create_column(&mut self.rng));
+        column_next.set_falling(x);
+
         self.column_falling = Some(column_next);
     }