+++ src/stage_handlers/ready.rs
@@ -1,19 +1,27 @@
+use std::time::Duration;
+
+use anyhow::Context;
 use ratatui::crossterm::event::{KeyCode, KeyEvent};
 
 use crate::{
     game::Game,
-    stage_handlers::{GameplayHandler, Stage, StageHandler},
+    stage_handlers::{FRAME_DURATION_IDLE, GameplayHandler, Stage, StageHandler},
 };
 
+#[derive(Debug)]
 pub struct ReadyHandler;
 
 impl StageHandler for ReadyHandler {
     fn handle_key_pressed_event(&mut self, game: &mut Game, key_event: KeyEvent) -> Option<Stage> {
         if key_event.code == KeyCode::Enter {
-            game.start();
+            game.start().context("Failed to start the game").ok()?;
             return Some(Stage::Gameplay(GameplayHandler::new()));
         }
         None
+    }
+
+    fn time_before_next_tick(&mut self, _game: &mut Game) -> Duration {
+        FRAME_DURATION_IDLE
     }
 
     fn update(&mut self, _game: &mut Game) -> Option<Stage> {